Add core spectrum object.
