+++ pymatgen/io/abinit/tasks.py
@@ -3941,7 +3941,7 @@ class BseTask(ManyBodyTask):
 
 class OpticTask(Task):
     """
-    Task for the computation of optical spectra with optic i.e.
+    Task for the computation of optical xas with optic i.e.
     RPA without local-field effects and velocity operator computed from DDK files.
     """
     color_rgb = np.array((255, 204, 102)) / 255
